+++ src/main/java/cn/wthee/pcrtool/adapters/CharacterLevelExpAdapter.kt
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import cn.wthee.pcrtool.MyApplication
 import cn.wthee.pcrtool.R
-import cn.wthee.pcrtool.database.entity.CharacterExperienceAll
+import cn.wthee.pcrtool.database.view.CharacterExperienceAll
 import cn.wthee.pcrtool.databinding.ItemLevelExpBinding
