+++ src/main/java/cn/wthee/pcrtool/database/AppDatabase.kt
@@ -65,11 +65,21 @@ abstract class AppDatabase : RoomDatabase() {
 
 
         private fun buildDatabase(): AppDatabase {
-            return Room.databaseBuilder(
-                MyApplication.getContext(),
-                AppDatabase::class.java,
-                DATABASE_Name
-            ).fallbackToDestructiveMigration().build()
+            try {
+                return Room.databaseBuilder(
+                    MyApplication.getContext(),
+                    AppDatabase::class.java,
+                    DATABASE_Name
+                ).fallbackToDestructiveMigration().build()
+            }catch (e : Exception){
+                //TODO mirgin
+                return Room.databaseBuilder(
+                    MyApplication.getContext(),
+                    AppDatabase::class.java,
+                    DATABASE_Name
+                ).fallbackToDestructiveMigration().build()
+            }
+
         }
     }
