+++ src/main/java/cn/wthee/pcrtool/database/entity/CharacterData.kt
@@ -28,52 +28,4 @@ data class CharacterData(
     @ColumnInfo(name = "only_disp_owned") var onlyDispOwned: Int,
     @ColumnInfo(name = "start_time") var startTime: String,
     @ColumnInfo(name = "end_time") var endTime: String,
-    //jp
-//    @ColumnInfo(name = "is_limited") val isLimited: Int,
-//    @ColumnInfo(name = "cutin1_star6") val cutin1Star6: Int,
-//    @ColumnInfo(name = "cutin2_star6") val cutin2Star6: Int
-) : Serializable {
-
-//    constructor(
-//        dataId: Int,
-//        name: String,
-//        kana: String,
-//        prefabId: Int,
-//        rarity: Int,
-//        motion_type: Int,
-//        seType: Int,
-//        moveSpeed: Int,
-//        position: Int,
-//        atkType: Int,
-//        atkTime: Double,
-//        cutin1: Int,
-//        cutin2: Int,
-//        guildId: Int,
-//        exSkillDisplay: Int,
-//        comment: String,
-//        onlyDispOwned: Int,
-//        startTime: String,
-//        endTime: String
-//    ) :
-//            this(
-//                dataId,
-//                name,
-//                kana,
-//                prefabId,
-//                rarity,
-//                motion_type,
-//                seType,
-//                moveSpeed,
-//                position,
-//                atkType,
-//                atkTime,
-//                cutin1,
-//                cutin2,
-//                guildId,
-//                exSkillDisplay,
-//                comment,
-//                onlyDispOwned,
-//                startTime,
-//                endTime,0,0,0
-//    )
-}
+) : Serializable
