+++ src/main/java/cn/wthee/pcrtool/database/entity/UniqueEquipmentEnhanceRate.kt
@@ -4,6 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import cn.wthee.pcrtool.database.view.Attr
 
 //专武装备提升
 @Entity(tableName = "unique_equipment_enhance_rate")
