+++ src/main/java/cn/wthee/pcrtool/database/entityjp/EquipmentDataJP.kt
@@ -4,7 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import cn.wthee.pcrtool.database.entity.Attr
+import cn.wthee.pcrtool.database.view.Attr
 import java.io.Serializable
 
 @Entity(tableName = "equipment_data")
