+++ src/main/java/cn/wthee/pcrtool/database/view/PvpCharacterData.kt
@@ -1,4 +1,4 @@
-package cn.wthee.pcrtool.database.entity
+package cn.wthee.pcrtool.database.view
 
 import androidx.room.ColumnInfo
 import com.google.gson.JsonArray
