+++ src/main/java/cn/wthee/pcrtool/database/view/UniqueEquipmentMaxData.kt
@@ -1,18 +1,17 @@
-package cn.wthee.pcrtool.database.entity
+package cn.wthee.pcrtool.database.view
 
 import androidx.room.ColumnInfo
 import androidx.room.Embedded
 import cn.wthee.pcrtool.utils.Constants.UNKNOW_EQUIP_ID
 import java.io.Serializable
 
-data class EquipmentMaxData(
+data class UniqueEquipmentMaxData(
+    @ColumnInfo(name = "unit_id") val unitId: Int,
     @ColumnInfo(name = "equipment_id") val equipmentId: Int,
     @ColumnInfo(name = "equipment_name") val equipmentName: String,
-    @ColumnInfo(name = "type") val type: String,
     @ColumnInfo(name = "description") val description: String,
-    @ColumnInfo(name = "craft_flg") val craftFlg: Int,
-    @ColumnInfo(name = "require_level") val requireLevel: Int,
-    @ColumnInfo(name = "promotion_level") val promotionLevel: Int,
+    @ColumnInfo(name = "max_level") val maxLevel: Int,
+    @ColumnInfo(name = "rank") val rank: Int,
     @Embedded val attr: Attr,
 ) : Serializable {
 
@@ -20,12 +19,11 @@ data class EquipmentMaxData(
 
     companion object {
         fun unknow() =
-            EquipmentMaxData(
+            UniqueEquipmentMaxData(
+                0,
                 UNKNOW_EQUIP_ID,
                 "？？？",
                 "",
-                "",
-                0,
                 0,
                 0,
                 Attr()
@@ -34,5 +32,3 @@ data class EquipmentMaxData(
     }
 
 }
-
-
