+++ src/main/java/cn/wthee/pcrtool/ui/detail/character/CharacterPromotionViewModel.kt
@@ -5,10 +5,10 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import cn.wthee.pcrtool.data.CharacterRepository
 import cn.wthee.pcrtool.data.EquipmentRepository
-import cn.wthee.pcrtool.database.entity.Attr
-import cn.wthee.pcrtool.database.entity.add
-import cn.wthee.pcrtool.database.entity.multiply
-import cn.wthee.pcrtool.database.entity.EquipmentMaxData
+import cn.wthee.pcrtool.database.view.Attr
+import cn.wthee.pcrtool.database.view.add
+import cn.wthee.pcrtool.database.view.multiply
+import cn.wthee.pcrtool.database.view.EquipmentMaxData
 import cn.wthee.pcrtool.utils.Constants.UNKNOW_EQUIP_ID
 import kotlinx.coroutines.launch
