+++ src/main/java/cn/wthee/pcrtool/ui/detail/equipment/EquipmentDetailsViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import cn.wthee.pcrtool.data.EquipmentRepository
-import cn.wthee.pcrtool.database.entity.EquipmentDropInfo
-import cn.wthee.pcrtool.database.entity.EquipmentMaterial
-import cn.wthee.pcrtool.database.entity.EquipmentMaxData
+import cn.wthee.pcrtool.database.view.EquipmentDropInfo
+import cn.wthee.pcrtool.database.view.EquipmentMaterial
+import cn.wthee.pcrtool.database.view.EquipmentMaxData
 import kotlinx.coroutines.launch
