+++ src/main/java/cn/wthee/pcrtool/ui/main/EquipmentViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import cn.wthee.pcrtool.data.EquipmentRepository
-import cn.wthee.pcrtool.database.entity.EquipmentMaxData
-import cn.wthee.pcrtool.database.entity.UniqueEquipmentMaxData
+import cn.wthee.pcrtool.database.view.EquipmentMaxData
+import cn.wthee.pcrtool.database.view.UniqueEquipmentMaxData
 import kotlinx.coroutines.launch
